+++ ktuvitDownloader/__main__.py
@@ -1,13 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 import ConfigParser
-import base64
 import json
 import logging
 import os
 import sys
 from datetime import datetime
-from getpass import getpass
 from logging import handlers
 from time import sleep
 
@@ -15,9 +13,9 @@ import yaml
 from appdirs import AppDirs
 from guessit import guessit
 
-from ktuvitDownloader.CustomExceptions import CantFindSubtitleException, WrongLoginException
+from ktuvitDownloader.CustomExceptions import CantFindSubtitleException
 from ktuvitDownloader.__version__ import __version__
-from ktuvitDownloader.connection import Connection
+from ktuvitDownloader.downloader import Downloader
 from ktuvitDownloader.const import *
 from ktuvitDownloader.files import get_paths_files, move_finshed
 from ktuvitDownloader.options import args_parse, parse_log
@@ -72,46 +70,13 @@ def main():
             options.dest_path = True
 
     if options.specific:
-        base_dir = options.specific
-        if not os.path.isdir(base_dir):
-            ext = os.path.splitext(base_dir)[1]
-            if ext not in VIDEO_EXT:
-                print "The path you supplied  is not known video format.\nPlease try again."
-                exit(-1)
-            else:
-                file_name = base_dir.split("\\")[-1]
-                base_dir = os.path.dirname(os.path.realpath(base_dir))
-                dest_dir = os.path.abspath(os.path.join(base_dir, os.pardir))
-                config.read(config_file)
-                downloader(base_dir, dest_dir, {os.path.join(base_dir, file_name): guessit(file_name)}, specific=True)
-                exit(0)
-        dest_dir = os.path.abspath(os.path.join(base_dir, os.pardir))
+        base_dir, dest_dir = handle_specific(options)
 
     if options.base_path:
-        base_dir = raw_input("Enter the path to base directory (where all your files are): ")
-        while not os.path.isdir(base_dir):
-            print "Can't find this directory."
-            base_dir = raw_input("Try again: ")
-        try:
-            config.add_section("Directories")
-        except ConfigParser.DuplicateSectionError:
-            pass
-        config.set("Directories", "base_dir", base_dir)
-        with open(config_file, "wb") as f:
-            config.write(f)
+        base_dir = handle_dir("base_dir")
 
     if options.dest_path:
-        dest_dir = raw_input("Enter the path to dest directory (where you want to move your files. it can be the "
-                             "same as base): ")
-        while not os.path.isdir(dest_dir):
-            print "Can't find this directory."
-            dest_dir = raw_input("Try again: ")
-        try:
-            config.set("Directories", "dest_dir", dest_dir)
-        except ConfigParser.DuplicateSectionError:
-            pass
-        with open(config_file, "wb") as f:
-            config.write(f)
+        dest_dir = handle_dir("dest_dir")
 
     config.read(config_file)
     if not options.specific:
@@ -125,7 +90,41 @@ def main():
         sleep(2)
         exit(1)
 
-    downloader(base_dir, dest_dir, files, options.specific)
+    download_handler(base_dir, dest_dir, files, options.specific)
+
+
+def handle_dir(dir_str):
+    dir = raw_input("Enter the path to : " + dir_str)
+    while not os.path.isdir(dir):
+        print "Can't find this directory."
+        dir = raw_input("Try again: ")
+    try:
+        config.set("Directories", dir_str, dir)
+    except ConfigParser.DuplicateSectionError:
+        pass
+    with open(config_file, "wb") as f:
+        config.write(f)
+    return dir
+
+
+def handle_specific(options):
+    base_dir = options.specific
+    if not os.path.isdir(base_dir):
+        ext = os.path.splitext(base_dir)[1]
+        if ext not in VIDEO_EXT:
+            print "The path you supplied  is not known video format.\nPlease try again."
+            exit(-1)
+        else:
+            file_name = base_dir.split("\\")[-1]
+            base_dir = os.path.dirname(os.path.realpath(base_dir))
+            dest_dir = os.path.abspath(os.path.join(base_dir, os.pardir))
+            config.read(config_file)
+            download_handler(base_dir, dest_dir, {
+                os.path.join(base_dir, file_name): guessit(file_name)
+            }, specific=True)
+            exit(0)
+    dest_dir = os.path.abspath(os.path.join(base_dir, os.pardir))
+    return base_dir, dest_dir
 
 
 def check_user_data_dir():
@@ -202,26 +201,32 @@ def get_handler(isFile):
     return handler
 
 
-def downloader(base_dir, dest_dir, files, specific=False):
+def download_handler(base_dir, dest_dir, files, specific=False):
     downloaded = []
-    con = Connection(app_dir)
+    downloader = Downloader(app_dir, logger)
     for path, data in files.items():
         try:
-            vid_ext = os.path.splitext(path)[1]
-            generic_path = os.path.splitext(path)[0]
-            con.download(path)
-            logger.info("Found " + generic_path + "!")
-            print "Found", generic_path.split("\\")[-1]
-            downloaded.append((generic_path, vid_ext))
+            one_download_handler(downloader, data, downloaded, path)
         except CantFindSubtitleException as e:
             logger.warn(e)
         except Exception as e:
             logger.error(repr(e))
+    downloader.close(specific)
     if move_finshed(downloaded, base_dir, dest_dir):
-        print "\n\nDone! check your Dest folder. you may find surprise\nTry again in a few hours, to prevent the " \
-              "chance you'll get ban"
+        print "\n\nDone! check your Dest folder. you may find a surprise..."
     else:
-        print "Done! try again in a few hours, to prevent the chance you'll get ban"
+        print "Done!"
+
+
+def one_download_handler(downloader, data, downloaded, path):
+    vid_ext = os.path.splitext(path)[1]
+    generic_path = os.path.splitext(path)[0]
+    sub_file, sub_ext = downloader.download(data, generic_path.split("\\")[-1])
+    logger.info("Found " + path + "!")
+    print "Found", path.split("\\")[-1]
+    with open(generic_path + sub_ext, "w") as f:
+        f.write(sub_file)
+    downloaded.append((generic_path, sub_ext, vid_ext))
 
 
 if __name__ == "__main__":